% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{R package translating DNA codons to amino acid letters}
\usage{
dna_codons_to_aa(codons)
}
\value{
None
}
\description{
Input DNA sequence codons and retrieve amino acid sequence in one-letter format
}
