% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{Get codons from a DNA string}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{dna}{a string of nucleotides (only DNA, not usable on RNA), 
which will be subset into codons.}

\item{s}{the starting position for extracting the codons. Default is set to 1.}
}
\value{
returns \code{codons}, the codons extracted from the originating DNA string.
}
\description{
\code{mk_codons} takes a DNA string and uses the \code{substring} 
function to divide it into 3-letter codons.
}
