% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Create a random DNA sequence based on the wanted length}
\usage{
random_dna(l)
}
\arguments{
\item{l}{is an integer representing the length of sequence}
}
\value{
returns a random DNA sequence of length \code{l}
}
\description{
\code{random_dna} takes a number as an input argument, which is the length of the 
DNA sequence, and returns a random sequence of that length.
}
