% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{Retrieve amino acid sequence from DNA codons #' 
\code{dna_codons_to_aa} takes DNA sequence codons and uses 
\code{std_code_table} to convert three letter DNA codons into a one letter #' amino acid sequence}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{codons}{is a string of codons in the format "CAA" "ACA" "TAG"}
}
\value{
a one-letter amino acid sequence where "*" denotes a stop codon
}
\description{
Retrieve amino acid sequence from DNA codons #' 
\code{dna_codons_to_aa} takes DNA sequence codons and uses 
\code{std_code_table} to convert three letter DNA codons into a one letter #' amino acid sequence
}
