% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Create a random DNA sequence based on the size wanted}
\usage{
random_dna(l)
}
\arguments{
\item{l, }{which is an integer representing the lenght of sequence}
}
\value{
a DNA sequence of length l
}
\description{
\code{random_dna} takes a number as an input argument, which is the sizeof the 
DNA sequence, and returns random sequence of that size containing four bases
}
